Update README.md to include general styling guidelines
### General Guidelines

Code should be clear, tidy, and concise where possible, with meaningful variable names and inline notes to explain code where necessary. Sections of a file should be indicated by meaningful section titles. Personal notes should not be included in project files, but should be written in a file labelled notes.txt. notes.txt will only be accessible on your personal computer and will not appear to other contributors or in the project codebase on GitHub.